refactor: log-detail 구조 개선
